pyproject.toml
[project]
name = "vision-ai-tools"
version = "0.1.0"
description = "A collection of computer-vision tools for LLM Agents (image/video processing & understanding)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "awmthink", email = "awmthink@outlook.com" },
]
license = { file = "LICENSE" }
keywords = ["computer-vision", "video", "image", "tools", "skills", "ai", "llm", "agent"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
    "boto3>=1.28.0",
    "numpy",
    "opencv-python-headless",
    "python-dotenv",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/awmthink/vision-ai-tools"
Source = "https://github.com/awmthink/vision-ai-tools"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vision_ai_tools"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py311"]
